Metadata file processing merged into base class
